Method invocation change
Changed the point at which in execution the method is invoked/called.

This aims at ensuring the method
MasterFragment#onFragmentViewCreated(view, savedInstanceState); used to
attach the created view(s) data is called before the method to create
the views is returned. Thus when the create method returns, it returns
with the attached data.